requirements.txt
django==3.0.2
djangorestframework==3.11.0
python-dateutil=2.8.1